added TodoItem component
