+++ src/api/todos.ts
@@ -9,4 +9,8 @@ export const deleteTodoById = (todoId: number) => {
   return client.delete(`/todos/${todoId}`);
 };
 
-// Add more methods here
+export const createTodo = (fieldsToCreate: Omit<Todo, 'id'>) => {
+  const body = JSON.stringify(fieldsToCreate);
+
+  return client.post<Todo>('/todos', body);
+};
