+++ src/components/Header/Header.tsx
@@ -5,22 +5,24 @@ interface Props {
   newTodoField: React.RefObject<HTMLInputElement>,
 }
 
-export const Header: React.FC<Props> = memo(({ newTodoField }) => (
-  <header className="todoapp__header">
-    <button
-      data-cy="ToggleAllButton"
-      type="button"
-      className="todoapp__toggle-all active"
-    />
-
-    <form>
-      <input
-        data-cy="NewTodoField"
-        type="text"
-        ref={newTodoField}
-        className="todoapp__new-todo"
-        placeholder="What needs to be done?"
+export const Header: React.FC<Props> = memo(({ newTodoField }) => {
+  return (
+    <header className="todoapp__header">
+      <button
+        data-cy="ToggleAllButton"
+        type="button"
+        className="todoapp__toggle-all active"
       />
-    </form>
-  </header>
-));
+
+      <form>
+        <input
+          data-cy="NewTodoField"
+          type="text"
+          ref={newTodoField}
+          className="todoapp__new-todo"
+          placeholder="What needs to be done?"
+        />
+      </form>
+    </header>
+  );
+});
